Convert all /api/[namespaceName]/[projectName] routes to Typescript
